+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `provideMemproofs` endpoint
 - `allowEmailPasswordAuth` authOpts fields
 - `hasMemproofs` field in `AgentState`
+- `websdk_version` is now passed with authOpts to chaperone
+
+### Removed
+- chaperone version check
 
 ## [0.6.19-prerelease] - 2024-03-26